# Improve performance of some archiving queries