@tsteur 2022-02-17
FYI I will keep monitoring how CPU changes based of this change and merge next week.