@tsteur 2022-02-17
So far it looks like the CPU decreases by 50%.